added csrf-token logout, added edit.js documentation, various CSS
