+++ public/assets/js/addFileName.js
@@ -1,6 +1,8 @@
-$('.custom-file-input').on('change', function (e) {
-    let inputFile = e.currentTarget;
-    alert(inputFile);
-    $(inputFile).parent().find('.custom-file-label').html(inputFile.files[0].name);
+$(document).ready(function () {
+    
+    $('.custom-file-input').on('change', function (e) {
+        let inputFile = e.currentTarget;
+        alert(inputFile);
+        $(inputFile).parent().find('.custom-file-label').html(inputFile.files[0].name);
+    });
 });
-
