+++ public/assets/js/loadMoreTricks.js
@@ -1,31 +1,27 @@
 window.onload = () => {
-    $(document).ready(function() {
-       $(".tricks-more").hide().slice(0, 5).css("display", "flex");
-    
-       $(".scroll-to-up").hide();
-        
-        $("#loadMore").click(function(e){ 
-            e.preventDefault();
+    $(".tricks-more").hide().slice(0, 5).css("display", "flex");
+
+    $(".scroll-to-up").hide();
+
+    $("#loadMore").click(function (e) {
+        e.preventDefault();
         $(".tricks-more:hidden").slice(0, 5).css("display", "flex");
-        if($(".tricks-more:hidden").length ===5){
-            $(".scroll-to-up").show(); 
-         }
-            if($(".tricks-more:hidden").length ===0){ 
-                $("#loadMore").attr("disabled", "disabled");
-                alert("No more!");
-            }
-        });
+        if ($(".tricks-more:hidden").length === 5) {
+            $(".scroll-to-up").show();
+        }
+        if ($(".tricks-more:hidden").length === 0) {
+            $("#loadMore").attr("disabled", "disabled");
+            alert("No more!");
+        }
+    });
 
-    })
 
-    $(document).ready(function() {
-        
-        $(".media-n").hide();
-        $(".bt-edit-poster").hide();
-        $('.see').click(function(){
-            $(".media-n").toggle();
-            $(".bt-edit-poster").toggle();
-            $(this).val( $(this).val() == 'Not See Medias' ? 'See Medias' : 'Not See Medias' );
-        });
-    })
-}
+    $(".media-n").hide();
+    $(".bt-edit-poster").hide();
+    $('.see').click(function () {
+        $(".media-n").toggle();
+        $(".bt-edit-poster").toggle();
+        $(this).val($(this).val() == 'Not See Medias' ? 'See Medias' : 'Not See Medias');
+    });
+
+};
