+++ templates/security/forgotten.html.twig
@@ -2,8 +2,15 @@
 
 {% block title %}Forgot Password{% endblock %}
 
-{% block body %}
-    <h1>Mot de passe oublié</h1>
-    {{ form(emailForm) }}
-
+{% block body %}<br><br><br>
+<h1 class="text-center my-5">Forgot your password</h1><br>
+<div class="container my-5 text-center">
+    <div class="border py-5 px-5">
+        {{ form_start(emailForm)}}
+        {{ form_widget(emailForm) }}
+        <button type="submit" class="btn-roun-shadow btn btn-primary">Send !</button>
+        {{ form_end(emailForm) }}
+    </div><br>
+    <br><br><br><br><br><br><br><br>
+</div>
 {% endblock %}
