+++ templates/security/passwordUpdate.html.twig
@@ -14,8 +14,8 @@
     {{ form_start(form) }}
     {{ form_widget(form) }}
 
-    <button type="submit" class="btn btn-primary">Update password</button>
-    
+    <button type="submit" class="btn-roun-shadow btn btn-primary">Update password</button>
+
     {{ form_end(form) }}
 </div>
