+++ templates/security/resetPassword.html.twig
@@ -1,20 +1,25 @@
 {% extends 'base.html.twig' %}
- 
+
 {% block title %}Nouveau mot de passe{% endblock %}
- 
+
 {% block body %}
-<h1>Réinitialisation du mot de passe</h1>
-<p>Veuillez entrer votre nouveau mot de passe ci-dessous.</p>
-<form method="post">
-    <div class="form-group">
-        <label for="_username">Username</label>
-        <input type="text" placeholder="Your email" required name="_username" class="form-control">
-    </div><br><br>
-    <div class="form-group">
-    <label for="inputPassword">Password</label>
-    <input type="password" name="password" id="inputPassword" placeholder="new password" required>
-    <input type="hidden" name="token" value="{{ token }}">
-    </div>
-    <button type="submit" id="valider">Validate password</button>   
-</form>
+<div class="container text-center"><br><br>
+    <h1 class="my-5">Réinitialisation du mot de passe</h1>
+    <p class="mb-5">Veuillez entrer votre nouveau mot de passe ci-dessous.</p>
+    <div class=" border py-4 px-5">
+        <form method="post">
+            <div class="form-group">
+                <label for="_username">Username</label>
+                <input type="text" placeholder="Your email" required name="_username" class="form-control">
+            </div><br><br>
+            <div class="form-group">
+                <label for="inputPassword">Password</label>
+                <input type="password" name="password" id="inputPassword" placeholder="new password" required
+                    class="form-control">
+                <input type="hidden" name="token" value="{{ token }}">
+            </div>
+            <button class="btn-roun-shadow btn btn-success my-5" type="submit" id="valider">Validate password</button>
+        </form>
+    </div><br><br><br>
+</div>
 {% endblock %}
