+++ templates/shared/_navbar.html.twig
@@ -56,8 +56,11 @@
           </div>
         </li>
         <li class="nav-item d-flex">
-          <a class="nav-link" href="{{ path('security_logout') }}">Logout</a>
+          <a class="nav-link" href="#" onclick="event.preventDefault(); document.getElementById('js-logout-form').submit();">Logout</a>
         </li>
+        <form id="js-logout-form" action="{{ path('security_logout') }}" method="POST" style="display:none;">
+          <input type="hidden" name="_csrf_token" value="{{ csrf_token('logout') }}">
+        </form>
         {% endif %}
     </div>
   </nav>
