--- templates/trick/_delete_comment_form.html.twig
@@ -1,7 +1,0 @@
-<form method="post" action="{{ path('comment_delete', {'id': comment.id}) }}" onsubmit="return confirm('Are you sure you want to delete this comment?');">
-    <input type="hidden" name="_method" value="DELETE">
-    <input type="hidden" name="_token" value="{{ csrf_token('delete' ~ comment.id) }}">
-    {% if is_granted('edit', trick) %}
-    <button class="py-0 px-2 btn btn-link text-blue"><i class="fas fa-trash-alt"></i></button>
-    {% endif %}
-</form>
