+++ templates/trick/_delete_form.html.twig
@@ -2,6 +2,6 @@
     <input type="hidden" name="_method" value="DELETE">
     <input type="hidden" name="_token" value="{{ csrf_token('delete' ~ trick.id) }}">
     {% if is_granted('edit', trick) %}
-    <button class="d-flex d-inline-block btn-delete-image btn btn-danger fas fa-trash-alt"></button>
+    <button class="btn-roun-shadow d-flex d-inline-block btn-delete-image btn btn-danger fas fa-trash-alt"></button>
     {% endif %}
 </form>
