+++ templates/trick/index.html.twig
@@ -11,7 +11,7 @@
 {% if app.user %}
 <div class= "d-flex justify-content-center">
 
-<a class="btn-rounded btn btn-secondary text-white nav-link shadow" href="{{ path('trick_new') }}">Create new Trick !</a>
+<a class="btn-roun-shadow btn-rounded btn btn-secondary text-white nav-link shadow" href="{{ path('trick_new') }}">Create new Trick !</a>
 </div><br><br><br>
 {% endif %}
 {% endblock %}
