+++ templates/trick/new.html.twig
@@ -2,16 +2,16 @@
 {% block title %}New Trick{% endblock %}
 {% form_theme form _self %}
 {% block body %}
-<div class="container mt-4 text-center"><br>
+<div class="container text-center"><br>
 	<h1>Create new Trick</h1>
-	<section class="add-trick">
+	<section class="add-trick my-4">
 		{{ include('trick/_form.html.twig') }}
 	</section>
 	<a href="{{ path('trick_index') }}">back to list</a>
 </div>
 {% endblock %}
 {% block form_row %}
-<div class="form-group">
+<div class="form-group my-4">
 	{#{{dump()}}#}
 	{{ form_label(form) }}
 	{{ form_widget(form) }}
@@ -21,7 +21,7 @@
 <h3>Add pictures</h3>
 {{ form_widget(form) }}
 <div class=>
-	<button type="button" class="btn btn-primary btn-add-images" id="add-images">Add Picture</button>
+	<button type="button" class="btn-roun-shadow btn btn-primary btn-add-images my-4" id="add-images">Add Picture</button>
 </div>
 {% endblock %}
 {% block _trick_images_entry_widget %}
@@ -35,7 +35,7 @@
 {% block _trick_videos_row %}
 <h3>Add videos</h3>
 {{ form_widget(form) }}
-<button type="button" class="btn btn-primary btn-add-videos" id="add_videos">Add video</button><br><br>
+<button type="button" class="btn-roun-shadow btn btn-primary btn-add-videos my-4" id="add_videos">Add video</button><br><br>
 {% endblock %}
 {% block _trick_videos_entry_widget %}
 <div id="block_{{id}}">
@@ -67,6 +67,6 @@
 <script src="https://stackpath.bootstrapcdn.com/bootstrap/4.3.1/js/bootstrap.min.js"
 	integrity="sha384-JjSmVgyd0p3pXB1rRibZUAYoIIy6OrQ6VrjIEaFf/nJGzIxFDsf4x0xIM+B07jRM"
 	crossorigin="anonymous"></script>
-<script src="{{ asset('assets/js/handleFormPics.js') }}"></script>
+<script src="{{ asset('assets/js/handleForm.js') }}"></script>
 <script src="{{ asset('assets/js/addFileName.js') }}"></script>
 {% endblock %}
