+++ templates/trick/tricksListTricks.html.twig
@@ -35,7 +35,7 @@
                         <div class="d-flex col-12 p-3">
                             {% if is_granted('edit', trick) %}
                             <a class="px-2" href="{{ path('trick_edit', {'slug': trick.slug}) }}"><i
-                                    class="btn btn-primary fas fa-pencil-alt"></i></a>
+                                    class="btn-roun-shadow btn btn-primary fas fa-pencil-alt"></i></a>
                             {{ include('trick/_delete_form.html.twig') }}
                             {% endif %}
                             <a href="{{ path('trick_like', {'slug': trick.slug}) }}" class="btn btn-link js_like">
@@ -54,7 +54,7 @@
     </div>
     {% endfor %}
     <div class="col-12 d-flex">
-        <button class="btn-rounded btn btn-secondary mx-auto" id="loadMore">Load More</button>
+        <button class="btn-roun-shadow btn-rounded btn btn-secondary mx-auto" id="loadMore">Load More</button>
     </div>
     {% block javascripts %}
     <script src="{{ asset('assets/js/pics.js') }}"></script>
