+++ templates/user/profile.html.twig
@@ -15,13 +15,13 @@
         {% endif %}
         <br>
         {% if user.tricks | length > 0 %}
-        <a href="{{ path('app_get_trick') }}"><span class="badge badge-primary">{{user.tricks | length}} Tricks published</span></a>
+        <a href="{{ path('app_get_trick') }}"><span class="badge badge-success badge-shadw">{{user.tricks | length}} Tricks published</span></a>
         {% else %}
-        <span class="badge badge-primary">You have no trick published</span>
+        <span class="badge badge-warning">You have no trick published</span>
         {% endif %}
         <br>
         {% if user.comments | length > 0 %}
-        <a href="{{ path('app_get_comment') }}"><span class="badge badge-primary">{{user.comments | length}} comments published</span></a>
+        <a href="{{ path('app_get_comment') }}"><span class="badge badge-success badge-shadw">{{user.comments | length}} comments published</span></a>
         {% else %}
         <span class="badge badge-primary">You have no comments published</span>
         {% endif %}
@@ -30,7 +30,7 @@
         <div class="card-header"><br>Pseudo : {{ app.user.pseudo }}<br><br></div>
         <div class="card-content"><br>Mail : {{ app.user.mail }}<br><br></div>
         <div class="card-footer"><br>
-            <a class="btn btn-primary" href="{{path('app_edit_profile')}}">Edit {{ app.user.pseudo }} profile</a><br><br>
+            <a class="btn-roun-shadow btn btn-primary" href="{{path('app_edit_profile')}}">Edit {{ app.user.pseudo }} profile</a><br><br>
         </div>
     </div><br><br><br><br><br><br>
 </div>
